rearrange some more functions in scripts
