+++ packages/scripts/src/helpers/kubernetes.ts
@@ -145,20 +145,6 @@ async function showAssets(tsPort: number) {
     }
 }
 
-export async function logTCPPorts(service: string) {
-    try {
-        const command = 'netstat -an | grep \'^tcp\' | awk \'{print $4}\' | tr ".:" " " | awk \'{print $NF}\' | sort -n | uniq | tr "\n" " "';
-        const subprocess = await execaCommand(command, { shell: true, reject: false });
-        const { stdout, stderr } = subprocess;
-
-        if (stderr) {
-            throw new Error(stderr);
-        }
-        signale.info(`TCP Ports currently in use when starting ${service}:\n ${stdout}`);
-    } catch (err) {
-        signale.error(`Execa command failed trying to log ports: ${err}`);
-    }
-}
 
 export async function deletePersistentVolumeClaim(searchHost: string) {
     try {
