+++ packages/scripts/src/helpers/test-runner/utils.ts
@@ -8,7 +8,8 @@ import { ExecEnv, exec } from '../exec.js';
 import { readPackageInfo } from '../packages.js';
 import { TestOptions, GroupedPackages, TestFramework, TestFrameworks } from './interfaces.js';
 import { PackageInfo } from '../interfaces.js';
-import { getServicesForSuite, getPackageManager, type ArgsMap } from '../misc.js';
+import { getServicesForSuite, type ArgsMap } from '../misc.js';
+import { getPackageManager } from '../package-manager.js';
 import config from '../config.js';
 import signale from '../signale.js';
 import {
